requirements.txt
aiogram==3.12.0
python-dotenv==1.0.1
aiosqlite==0.20.0
SQLAlchemy==2.0.36
alembic==1.14.0
greenlet==3.1.1
